#76: Move SpringLoggingContextHelper under config package.
