+++ src/main/java/fi/fmi/avi/archiver/config/ParserConfig.java
@@ -15,7 +15,6 @@ import fi.fmi.avi.archiver.file.FileMetadata;
 import fi.fmi.avi.archiver.file.FileParser;
 import fi.fmi.avi.archiver.file.InputAviationMessage;
 import fi.fmi.avi.archiver.logging.LoggingContext;
-import fi.fmi.avi.archiver.logging.SpringLoggingContextHelper;
 import fi.fmi.avi.converter.AviMessageConverter;
 
 @Configuration
