+++ src/main/java/fi/fmi/avi/archiver/config/SpringLoggingContextHelper.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.archiver.logging;
+package fi.fmi.avi.archiver.config;
 
 import static java.util.Objects.requireNonNull;
 
@@ -13,6 +13,8 @@ import org.springframework.integration.handler.GenericHandler;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageHeaders;
 
+import fi.fmi.avi.archiver.logging.LoggingContext;
+import fi.fmi.avi.archiver.logging.NoOpLoggingContext;
 import fi.fmi.avi.archiver.spring.integration.dsl.ServiceActivators;
 
 public final class SpringLoggingContextHelper {
